+++ examples/react-nextjs/.next/server/chunks/455.js
@@ -1,1 +1,0 @@
-exports.id=455,exports.ids=[455],exports.modules={2348:()=>{}};