+++ examples/react-nextjs/.next/server/chunks/760.js
@@ -1,1 +1,0 @@
-exports.id=760,exports.ids=[760],exports.modules={760:()=>{}};